--- 1.1.1/2.1.3/2.1.3.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "self:2.1.3.xcodeproj">
-   </FileRef>
-</Workspace>
