--- 1.1.1/2.1.3/2.1.3.xcodeproj/xcuserdata/shotakawasaki.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "D770E7F0-84D6-4109-AB7D-FF0997CCE8D6"
-   type = "1"
-   version = "2.0">
-</Bucket>
