--- 1.1.1/2.1.3/2.1.3/Assets.xcassets/Image-4.imageset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "filename" : "lO46%btZSoS72EUdZukXVw_thumb_3-1.jpg",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "lO46%btZSoS72EUdZukXVw_thumb_3.jpg",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "lO46%btZSoS72EUdZukXVw_thumb_3-2.jpg",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
