+++ 2.1.1/2.1.1.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "self:2.1.1.xcodeproj">
+   </FileRef>
+</Workspace>
