+++ 2.1.1/2.1.1/Assets.xcassets/AppIcon.appiconset/Contents.json
