+++ 2.1.1/2.1.1/ViewController.swift
@@ -1,0 +1,30 @@
+//
+//  ViewController.swift
+//  2.1.1
+//
+//  Created by Shota on 2019/10/07.
+//  Copyright © 2019 Shota Kawasaki. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+    
+   
+    
+    @IBOutlet weak var startLabel: UILabel!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+       
+    
+        
+       
+
+        
+        // Do any additional setup after loading the view.
+    }
+   
+    
+}
+
