+++ 2.1.1/2.1.1/en.lproj/Localizable.strings
@@ -1,0 +1,9 @@
+/* 
+  Localizable.strings
+  2.1.1
+
+  Created by Shota on 2019/10/07.
+  Copyright © 2019 Shota Kawasaki. All rights reserved.
+*/
+
+"start" = "Started Swift" ;
