+++ 2.1.1/2.1.1/en.lproj/Main.strings
@@ -1,0 +1,3 @@
+
+/* Class = "UILabel"; text = "start"; ObjectID = "2ym-zq-6Ud"; */
+"2ym-zq-6Ud.text" = "Startrd Swift";
