+++ 2.1.1/2.1.1/ja.lproj/Localizable.strings
@@ -1,0 +1,10 @@
+/* 
+  Localizable.strings
+  2.1.1
+
+  Created by Shota on 2019/10/07.
+  Copyright © 2019 Shota Kawasaki. All rights reserved.
+*/
+
+"start" = "Swift講座をスタートしました。";
+
