+++ 2.1.1/2.1.1/ja.lproj/Main.strings
@@ -1,0 +1,3 @@
+
+/* Class = "UILabel"; text = "start"; ObjectID = "2ym-zq-6Ud"; */
+"2ym-zq-6Ud.text" = "swift講座をスタートしました";
